+++ backend/validators/userValidator.js
@@ -97,100 +97,133 @@ const validateRegisterUser = [  //validate user & email
     }
 ]
 
-/*const validateUpdateUser = [  //validate user unique & email unique
-    check('firstName')
-    .exists()
-    .not()
-    .isEmpty()
-    .withMessage('First name input is empty'),
+/*function validationUpdateUser (req, res, next){
+    const validateUpdateUser = [ ]
 
-    check('lastName')
-    .exists()
-    .not()
-    .isEmpty()
-    .withMessage('Last name input is empty'),
-
-    check('email')
-        .exists()
-        .isEmail()
-        .withMessage('Input is not an email')/////////////////////////////////////////////////////////////////////////////////////
-        .custom(async (value, { res }) => {
-            let email_resp;
-
-            let query = {
-                email: value
-            }
-
-            try {
-                email_resp = await User.findOne(query);
-            } catch (error) {
-                console.log(error);
-                return res.status(500).json({ ////////////////////////////////////////////////////////////////////////////////////////
-                    error: "Something went wrong while trying to connect to the database"
-                });
-            }
-
-            if (email_resp) {
-                console.log(error);
-                return res.status(500).json({ ////////////////////////////////////////////////////////////////////////////////////////
-                    error: "This email is already in use"
-                });
-            }
-        })
-        .withMessage('This email is already in use'),//////////////////////////////////////////////////////////////////////////////
-
-    check('user')
-        .exists()
-        .not()
-        .isEmpty()
-        .withMessage('Username input is empty')/////////////////////////////////////////////////////////////////////////////////////
-        .custom(async (value, { res }) => {
-            let user_resp;
-
-            let query = {
-                user: value
-            }
-
-            try {
-                user_resp = await User.findOne(query);
-            } catch (error) {
-                console.log(error);
-                return res.status(500).json({ ////////////////////////////////////////////////////////////////////////////////////////
-                    error: "Something went wrong while trying to connect to the database"
-                });
-            }
-
-            if (user_resp) {
-                console.log(error);
-                return res.status(500).json({ ////////////////////////////////////////////////////////////////////////////////////////
-                    error: "This user is already in use"
-                });
-            }
-        })
-        .withMessage('This user is already in use'), //////////////////////////////////////////////////////////////////////////////
-        
-    check('gender')
-        .exists()
-        .not()
-        .isEmpty()
-        .withMessage('Gender input is empty'),  
-    
-    check('password')
-        .exists()
-        .not()
-        .isEmpty()
-        .withMessage('Password input is empty'),
-    
-    check('role')
-        .exists()
-        .not()
-        .isEmpty()
-        .withMessage('Role input is empty'),
-
-    (req, res, next) => {
-        validateResult(req, res, next)
+    if (req.body.firstName) {
+        validateUpdateUser.push(
+            check('firstName')
+            .exists()
+            .not()
+            .isEmpty()
+            .withMessage('First name input is empty')
+        );
     }
-]*/
 
-module.exports = { validateRegisterUser }
-//module.exports = { validateUpdateUser }
+    if (req.body.lastName) {
+        validateUpdateUser.push(
+            check('lastName')
+            .exists()
+            .not()
+            .isEmpty()
+            .withMessage('Last name input is empty')
+        );
+    }
+
+    if (req.body.email) {
+        validateUpdateUser.push(
+            check(email)
+                .exists()
+                .isEmail()
+                .withMessage('Input is not an email')/////////////////////////////////////////////////////////////////////////////////////
+                .custom(async (value, { res }) => {
+                    let email_resp;
+            
+                    let query = {
+                        email: value
+                    }
+            
+                    try {
+                        email_resp = await User.findOne(query);
+                    } catch (error) {
+                        console.log(error);
+                        return res.status(500).json({ ////////////////////////////////////////////////////////////////////////////////////////
+                            error: "Something went wrong while trying to connect to the database"
+                        });
+                    }
+            
+                    if (email_resp) {
+                        console.log(error);
+                        return res.status(500).json({ ////////////////////////////////////////////////////////////////////////////////////////
+                            error: "This email is already in use"
+                        });
+                    }
+                })
+                .withMessage('This email is already in use'),//////////////////////////////////////////////////////////////////////////////
+        );
+    }
+
+    if (req.body.user) {
+        validateUpdateUser.push(
+            check('user')
+                .exists()
+                .not()
+                .isEmpty()
+                .withMessage('Username input is empty')/////////////////////////////////////////////////////////////////////////////////////
+                .custom(async (value, { res }) => {
+                    let user_resp;
+            
+                    let query = {
+                        user: value
+                    }
+            
+                    try {
+                        user_resp = await User.findOne(query);
+                    } catch (error) {
+                        console.log(error);
+                        return res.status(500).json({ ////////////////////////////////////////////////////////////////////////////////////////
+                            error: "Something went wrong while trying to connect to the database"
+                        });
+                    }
+            
+                    if (user_resp) {
+                        console.log(error);
+                        return res.status(500).json({ ////////////////////////////////////////////////////////////////////////////////////////
+                            error: "This user is already in use"
+                        });
+                    }
+                })
+                .withMessage('This user is already in use'), //////////////////////////////////////////////////////////////////////////////
+        );
+    }
+
+    if (req.body.gender) {
+        validateUpdateUser.push(
+            check('gender')
+            .exists()
+            .not()
+            .isEmpty()
+            .withMessage('Gender input is empty')
+        );
+    }
+
+    if (req.body.password) {
+        validateUpdateUser.push(
+            check('password')
+            .exists()
+            .not()
+            .isEmpty()
+            .withMessage('Password input is empty')
+        );
+    }
+
+    if (req.body.role) {
+        validateUpdateUser.push(
+            check('role')
+            .exists()
+            .not()
+            .isEmpty()
+            .withMessage('Role input is empty')
+        );
+    }          
+
+    validateUpdateUser.push(
+        (req, res, next) => {
+            validateResult(req, res, next)
+        }
+    )
+
+    return validateUpdateUser;
+}*/
+
+module.exports = { validateRegisterUser/*, validationUpdateUser*/ }
